Repository refactor (no need to split for educational purposes)
Header files comment now shows the TechTalk title
