+++ Examples/GlobantPlus/App/GlobantPlus/Application/GlobantPlusApp.swift
@@ -1,6 +1,6 @@
 //
 //  PremiereApp.swift
-//  Premiere
+//  GlobantPlus
 //
 //  Created by Adolfo Vera Blasco on 3/1/23.
 //
