+++ Examples/GlobantPlus/App/GlobantPlus/Data/Queues/AmazonSQS/SQSActivityQueue.swift
@@ -1,6 +1,6 @@
 //
 //  SQSActivityQueue.swift
-//  Premiere
+//  GlobantPlus
 //
 //  Created by Adolfo Vera Blasco on 12/3/23.
 //
