+++ Examples/GlobantPlus/App/GlobantPlus/Domain/Model/GlobantPlusError.swift
@@ -1,6 +1,6 @@
 //
 //  PremiereError.swift
-//  Premiere
+//  GlobantPlus
 //
 //  Created by Adolfo Vera Blasco on 11/3/23.
 //
