+++ Examples/GlobantPlus/App/GlobantPlus/Domain/Model/Show.swift
@@ -1,6 +1,6 @@
 //
 //  Show.swift
-//  Premiere
+//  GlobantPlus
 //
 //  Created by Adolfo Vera Blasco on 6/3/23.
 //
