+++ Examples/GlobantPlus/App/GlobantPlus/Domain/Model/TrendingItem.swift
@@ -1,6 +1,6 @@
 //
 //  TrendingItem.swift
-//  Premiere
+//  GlobantPlus
 //
 //  Created by Adolfo Vera Blasco on 4/1/23.
 //
