--- Examples/GlobantPlus/App/GlobantPlus/Domain/Repositories/Dashboard/TrendingMovieListRepository.swift
@@ -1,5 +1,0 @@
-import Foundation
-
-protocol TrendingMovieListRepository {
-    func fetchTrendingMovies() async throws -> [TrendingItem]
-}
