--- Examples/GlobantPlus/App/GlobantPlus/Domain/Repositories/Dashboard/TrendingShowListRepository.swift
@@ -1,12 +1,0 @@
-//
-//  TrendingShowListRepository.swift
-//  GlobantPlus
-//
-//  Created by Adolfo Vera Blasco on 16/3/23.
-//
-
-import Foundation
-
-protocol TrendingShowListRepository {
-    func fetchTrendingShows() async throws -> [TrendingItem]
-}
