--- Examples/GlobantPlus/App/GlobantPlus/Domain/Repositories/Favorires/DeleteFavoriteRepository.swift
@@ -1,12 +1,0 @@
-//
-//  DeleteFavoriteRepository.swift
-//  Premiere
-//
-//  Created by Adolfo Vera Blasco on 11/3/23.
-//
-
-import Foundation
-
-protocol DeleteFavoriteRepository {
-    func delete(showId: ShowID) throws
-}
