--- Examples/GlobantPlus/App/GlobantPlus/Domain/Repositories/Favorires/FavoriteTrainingPOSTRepository.swift
@@ -1,12 +1,0 @@
-//
-//  NetworkFavoriteRepository.swift
-//  GlobantPlus
-//
-//  Created by Adolfo Vera Blasco on 14/3/23.
-//
-
-import Foundation
-
-protocol FavoriteTrainingPOSTRepository {
-    func post(mediaID: Int, forUser userID: String) async throws
-}
