--- Examples/GlobantPlus/App/GlobantPlus/Domain/Repositories/Favorires/InsertFavoriteRepository.swift
@@ -1,12 +1,0 @@
-//
-//  InsertFavoriteRepository.swift
-//  Premiere
-//
-//  Created by Adolfo Vera Blasco on 11/3/23.
-//
-
-import Foundation
-
-protocol InsertFavoriteRepository {
-    func insert(showId: ShowID)
-}
