+++ Examples/GlobantPlus/App/GlobantPlus/Domain/UseCases/Dashboard/DefaultTrendingShowListUseCase.swift
@@ -1,6 +1,6 @@
 //
 //  DefaultDashboardUseCase.swift
-//  Premiere
+//  GlobantPlus
 //
 //  Created by Adolfo Vera Blasco on 4/1/23.
 //
@@ -8,7 +8,7 @@
 import Foundation
 
 final class DefaultTrendingShowListUseCase: TrendingShowListUseCase {
-    let trendingShowListRepository: TrendingShowListRepository = TMDBTrendingShowListRepository()
+    let trendingShowListRepository: DashboardRepository = TMDBDashboardRepository()
     
     func fetchTrendingShows() async throws -> [TrendingItem] {
         return try await trendingShowListRepository.fetchTrendingShows()
