+++ Examples/GlobantPlus/App/GlobantPlus/Domain/UseCases/Favorites/DefaultFavoriteUseCase.swift
@@ -1,6 +1,6 @@
 //
 //  DefaultFavoriteUseCase.swift
-//  Premiere
+//  GlobantPlus
 //
 //  Created by Adolfo Vera Blasco on 11/3/23.
 //
@@ -8,24 +8,21 @@
 import Foundation
 
 final class DefaultFavoriteUseCase: FavoriteUseCase {  
-    private let insertRepository: InsertFavoriteRepository = CoreDataInsertFavoriteRepository()
-    private let deleteRepository: DeleteFavoriteRepository = CoreDataDeleteFavoriteRepository()
-    
-    private let favoriteTrainingPOST = AmazonFavoriteTrainingPOSTRepository()
-    private let favoriteTrainingDELETE = AmazonFavoriteTrainingDELETERepository()
+    private let favoriteRepository: FavoriteRepository = CoreDataFavoriteRepository()
+	private let favoriteTrainingRepository: FavoriteTrainingRepository = AmazonFavoriteTrainingRepository()
     
     func setFavorite(to newState: Bool, for show: ShowID) throws {
         if newState {
-            insertRepository.insert(showId: show)
+			favoriteRepository.insert(showId: show)
             
             Task(priority: .background) {
-                try? await favoriteTrainingPOST.post(mediaID: show, forUser: "Adolfo")
+                try? await favoriteTrainingRepository.post(mediaID: show, forUser: "Adolfo")
             }
         } else {
-            try deleteRepository.delete(showId: show)
+            try favoriteRepository.delete(showId: show)
             
             Task(priority: .background) {
-                try? await favoriteTrainingDELETE.delete(mediaID: show, forUser: "Adolfo")
+                try? await favoriteTrainingRepository.delete(mediaID: show, forUser: "Adolfo")
             }
         }
     }
