+++ Examples/GlobantPlus/App/GlobantPlus/Domain/UseCases/Show/DefaultShowUseCase.swift
@@ -1,6 +1,6 @@
 //
 //  DefaultShowUseCase.swift
-//  Premiere
+//  GlobantPlus
 //
 //  Created by Adolfo Vera Blasco on 6/3/23.
 //
