+++ Examples/GlobantPlus/App/GlobantPlus/Presentation/Scenes/Dashboard/Model/DashboardView+Model.swift
@@ -1,6 +1,6 @@
 //
 //  DashboardView+Model.swift
-//  Premiere
+//  GlobantPlus
 //
 //  Created by Adolfo Vera Blasco on 4/1/23.
 //
