+++ Examples/GlobantPlus/App/GlobantPlus/Presentation/Scenes/Dashboard/View/DashboardView.swift
@@ -1,6 +1,6 @@
 //
 //  DashboardView.swift
-//  Premiere
+//  GlobantPlus
 //
 //  Created by Adolfo Vera Blasco on 4/1/23.
 //
