+++ Examples/GlobantPlus/App/GlobantPlus/Presentation/Scenes/Show/Model/ShowView+Model.swift
@@ -1,6 +1,6 @@
 //
 //  ShowView+Model.swift
-//  Premiere
+//  GlobantPlus
 //
 //  Created by Adolfo Vera Blasco on 6/3/23.
 //
