+++ Examples/GlobantPlus/App/GlobantPlus/Presentation/Scenes/Show/View/ShowView.swift
@@ -1,6 +1,6 @@
 //
 //  ShowView.swift
-//  Premiere
+//  GlobantPlus
 //
 //  Created by Adolfo Vera Blasco on 18/1/23.
 //
