+++ Examples/GlobantPlus/App/GlobantPlus/Presentation/Scenes/Show/ViewModel/DefaultShowViewModel.swift
@@ -1,6 +1,6 @@
 //
 //  DefaultShowViewModel.swift
-//  Premiere
+//  GlobantPlus
 //
 //  Created by Adolfo Vera Blasco on 6/3/23.
 //
