+++ Examples/GlobantPlus/App/GlobantPlus/Presentation/Scenes/Show/ViewModel/ShowViewModel.swift
@@ -1,6 +1,6 @@
 //
 //  ShowViewModel.swift
-//  Premiere
+//  GlobantPlus
 //
 //  Created by Adolfo Vera Blasco on 6/3/23.
 //
